Create IsCharactersCanFormPalindrome.java
Anagram Palindrome
      
Given a string, Check if characters of the given string can be rearranged to form a palindrome. 
For example characters of “geeksogeeks” can be rearranged to form a palindrome “geeksoskeeg”, but characters of “geeksforgeeks” cannot be rearranged to form a palindrome.

Input:
First line consists of T test cases. Only line of test cases consists of String.

Output:
Single line output, print "Yes" if is possible to make it a palindrome else "No".

Constraints:
1<=T<=100
1<=|String length|<=1000

Example:
Input:
2
geeksogeeks
geeksforgeeks
Output:
Yes
No